Renamed the package to rhda. Followed suggestions from Jenkins team for publishing to update the pom.xml.
