+++ src/main/java/redhat/jenkins/plugins/rhda/service/PackageManagerService.java
@@ -1,0 +1,14 @@
+package redhat.jenkins.plugins.rhda.service;
+
+import java.io.IOException;
+import java.nio.file.Path;
+
+public interface PackageManagerService {
+
+    String getName();
+
+    String readFile(Path path) throws IOException;
+
+    String generateSbom(Path path) throws IOException;
+
+}
