+++ src/main/resources/redhat/jenkins/plugins/rhda/step/CRDAStep/config.jelly
@@ -1,0 +1,21 @@
+<?jelly escape-by-default='true'?>
+<!--
+  ~ Copyright © 2021 Red Hat Inc.
+	~
+	~ Licensed under the Apache License, Version 2.0 (the "License");
+	~ you may not use this file except in compliance with the License.
+	~ You may obtain a copy of the License at
+	~
+	~     http://www.apache.org/licenses/LICENSE-2.0
+	~
+	~ Unless required by applicable law or agreed to in writing, software
+	~ distributed under the License is distributed on an "AS IS" BASIS,
+	~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	~ See the License for the specific language governing permissions and
+	~ limitations under the License.
+	~ Author: Yusuf Zainee <yzainee@redhat.com>
+  ~
+  -->
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler">
+    <st:include class="redhat.jenkins.plugins.rhda.task.CRDABuilder" page="config.jelly" />
+</j:jelly>
