+++ src/main/resources/redhat/jenkins/plugins/rhda/task/CRDABuilder/help-consentTelemetry.html
